Changing attributes to String
Changing the type of JSON attributes from Integer to Long and String; Inclusion of addressId in interfaces

+++ customermanagement/src/main/java/br/com/ordertech/customermanagement/external/db/CustomerRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CustomerRepository extends JpaRepository<CustomerModel, Integer> {
+public interface CustomerRepository extends JpaRepository<CustomerModel, Long> {
     List<CustomerModel> findByName(String name);
 }
